Trekk ut delt oppsett av app i hook
