+++ packages/internarbeidsflate-decorator-v3/src/index.ts
@@ -1,4 +1,6 @@
-import App from './App';
 import Navspa from '@navikt/navspa';
+import Decorator from './Decorator';
+import LandingPage from './LandingPage';
 
-Navspa.eksporter('internarbeidsflate-decorator-v3', App);
+Navspa.eksporter('internarbeidsflate-decorator-v3', Decorator);
+Navspa.eksporter('internarbeidsflate-decorator-v3-fullscreen', LandingPage);
