+++ tsconfig.json
@@ -19,5 +19,5 @@
     "noFallthroughCasesInSwitch": true,
     "esModuleInterop": true
   },
-  "include": ["packages/internarbeidsflate-decorator-v3/src", "packages/internarbeidsflate-decorator-v3/App.tsx"],
+  "include": ["packages/internarbeidsflate-decorator-v3/src", "packages/internarbeidsflate-decorator-v3/Decorator.tsx", "packages/internarbeidsflate-decorator-v3/LandingPage.tsx"]
 }
